+++ src/pat/switch/_switch.scss
@@ -1,9 +1,9 @@
 li.mode-Red {
-    color: red !important;
+  color: red !important;
 }
 li.mode-Green {
-    color: green;
+  color: green;
 }
 li.mode-Blue {
-    color: blue;
+  color: blue;
 }
